backend/requirements.txt
Flask>=2.0
flask-cors>=3.0
joblib>=1.0
pandas>=1.0
numpy>=1.21
scikit-learn>=1.0
pymongo>=4.0
dnspython>=2.0
python-dotenv>=1.0.0
